requirements.txt
flask
flask-login
flask-wtf
flask-bcrypt
flask-sqlalchemy
mysql-connector-python
python-dotenv
weasyprint
